105 add the feature to update an expense attachment

+++ app/controllers/userController.js
@@ -292,16 +292,20 @@ const getUserById = async (req, res) => {
 const forgotPassword = async (req, res) => {
     try {
         const { email } = req.body;
+
+        const user = await User.findOne({ email });
+        if (user && user.isGoogle) {
+            return res.status(403).json({ error: LOCALE.googleUserCannotResetPassword });
+        }
+
         const existingEntry = await ForgotPassword.findOne({ email });
 
         if (existingEntry) {
             await ForgotPassword.findOneAndDelete({ email });
         }
 
-        const user = await User.findOne({ email });
-
         if (!user) {
-            return res.status(404).json({ error: LOCALE.userNotFound });
+            return res.status(404).json({ error: LOCALE.emailDoesNotBelongToUser });
         }
 
         const verificationCode = Math.floor(100000 + Math.random() * 900000); // 6-digit code
